# api/swagger/TmsSwagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: TMSAppConnector
  description: >-
    TMSAppConnector manages ISV integrations
  contact:
    email: maheshkumar.g@varsitymgmt.com
produces:
  - application/json
basePath: /api/v1
schemes:
- "https"
...
securityDefinitions:
       Bearer:
        type: apiKey
        name: Authorization
        in: header
paths:
...
  /assignStudentsForHomeworkView:
    get:
      summary: Teacher will Assign Homework to List of Students. 
      tags:
        - TeacherHomework
      operationId: assignStudentsForHomeworkView
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
       - in: query
         name: homeworkId
         type: integer
         required: true
        
      responses:
        200:
          description: Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/studentsHomeworkView'
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server error
          schema:
            $ref: '#/definitions/Error'
      security:
        - Bearer: []  
...
  /teacherQuestionLibrary:
    post:
      summary: Assign Students For Homework
      tags:
        - CreateOwnLibrary
      operationId: teacherQuestionLibrary
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "teacherQuestionLibrary"
        required: true
        schema:
          $ref: "#/definitions/teacherQuestionLibraryInfo"
      responses:
        200:
          description: "Successfully Created Library"
          schema:
             $ref: '#/definitions/Response'
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server error
          schema:
            $ref: '#/definitions/Error'
      security:
        - Bearer: []
  /getTeacherQuestionLibrary:
    get:
      summary: getTeacherQuestionLibrary
      operationId: getTeacherQuestionLibrary
      tags:
        - CreateOwnLibrary
      parameters:
        - in: query
          name: teacherId
          type: "string"
          required: true
      responses:
        200:
          description: Ok
          schema:
            type: array
            items:
              $ref: "#/definitions/teacherQuestionLibraryInfo"
      security:
        - Bearer: []
  /updateTeacherQuestionLibrary:   
    patch:
       tags:
       - CreateOwnLibrary
       summary: "To update QuestionLibrarys"
       operationId: "updateTeacherQuestionLibrary"
       produces:
       - "application/json"
       parameters:
       - in: body
         name: "updateTeacherQuestionLibrary"
         required: true
         description: "To updateTeacherQuestionLibrary"
         schema:
           $ref: "#/definitions/getTeacherQuestionLibraryInfo"
       security:
       - Bearer: []    
       responses:
         "200":
           description: "Successful operation"
           schema:
              $ref: "#/definitions/Response"
         "404":
           description: "Question details not available"   
...
definitions:
...
  Response:
    type: "object"
    properties:
      success:
          type: string
      message:
          type: string
...
  studentsHomeworkView:
      type: object
      properties:
      
       assignStudents:
        type: array
        description: Homework view(students,About,Questions)

        items:
          title: assignStudents
          description: students for homework assign
          type: object
          properties:
            nameAndAdmNo:
              type: string
              description: nameAndAdmNo
            admNo:
              type: string
              description: Adm No
            name:
              type: string
              description: Name  
            submissionDate:
              type: string
            status:
              type: string
            evaluationStatus:
              type: string  
            marksObtained:
               type: number
               format: float
       studentBranchs:
             type: array
             description: students branches for homework
        
             items:
              title: studentBranch
              description: student Branch for homework create
              type: object
              properties:
                branchId:
                 type: integer
                 format: int64
                 description: Student Branch Id
                branchName:
                 type: string
                 description: Student Branch Name
       studentClasses:
             type: array
             description: students classes for homework
        
             items:
              title: studentClass
              description: student class for homework create
              type: object
              properties:
               classId:
                type: integer
                format: int64
                description: Student Class Id
               className:
                type: string
                description: Student Class Name
       studentSections:
             type: array
             description: students classes for homework
        
             items:
              title: studentSetionInfo
              description: student section for homework create
              type: object
              properties:
               sectionId:
                type: integer
                format: int64
                description: Student Section Id
               sectionName:
                type: string
                description: Student Section Name          
       studentBatches:
             type: array
             description: students batches for homework
             items:
              title: studentAssignBatch
              description: student batches for homework create
              type: object
              properties:
               batchId:
                type: integer
                format: int64
                description: Student batch Id
               batchName:
                type: string
                description: Student batch Name
       studentOrientaion:
             type: array
             description: students Orientation for homework
             items:
              title: studentOrientaion
              description: student Orientation for homework create
              type: object
              properties:
               orientaionId:
                type: integer
                format: int64
                description: Student Orientation Id
               orientationName:
                type: string
                description: Student Orientation Name
       createdby:
         type: string
         description: created by
       createdOn:
         type: string
         description: created on  
       homeWorkId:
         type: integer
         format: int64
         description: homeWorkId
       homeWorkName:
         type: string
         description: homeWorkName  
       homeWorkType:
         type: string
         description: homeWorkTyp
       homeWorkStatus:
         type: string
         description: homeWorkStatus
       homeWorkSubject:
         type: string
         description: homeWorkSubject   
      
       description:
         type: string
         description: homework description
       startDate:
         type: string
         description: homework startDate
       endDate:
         type: string
         description: homework endDate

       assignQuestions:
        type: array
        description: assignment question files

        items:
          title: assignQuestions
          description: assignment question files
          type: object
          properties:
            assignmentFiles:
              type: string
              description: assignment Files
            name:
              type: string
              description: assignment name
            fileSize:
              type: number
              description: assignment file Size
            fileType:
              type: string
              description: assignment File Type
       
       assignmentTestPapers:

         type: array
         description: assignment question files
         items:
              $ref: "#/definitions/assignmentTestPaper"
...
  sectionAssignmentTest:
    type: object
    properties:
      subject:
        type: string
        description: subject
      questiontext:
        type: string
        description: Question Text
      choiceData1:
        type: string
        description: Choice Data1
      choiceData2:
        type: string
        description: Choice Data2
      choiceData3:
        type: string
        description: Choice Data3
      choiceData4:
        type: string
        description: Choice Data4
      solutionData:
        type: string
        description: Choice Data4
      questiontype:
        type: string
        description: Question type
      difficultylevel:
        type: string
        description: Difficulty Level
      objective:
        type: string
        description: Objective
      taxonomy:
        type: string
        description: Taxonomy
      relevance:
        type: string
        description: Relevance
      conceptlevel:
        type: string
        description: Conceptlevel
      syllabus:
        type: string
        description: syllabus
      topic:
        type: string
        description: Topic
      subtopic:
        type: string
        description: Subtopic
      branchId:
        type: string
        description: BranchId
      userId:
        type: string
        description: UserId
      stateId:
        type: string
        description: StateId
      status:
        type: string
        description: Status
      approverId:
        type: string
        description: ApproverId
      courses:
        type: string
        description: Courses
      board:
        type: string
        description: Board
      classes:
        type: string
        description: Classes
      chapter:
        type: string
        description: Chapterl
      questionanswer:
        type: string
        description: Questionanswer
      totalmarks:
        type: string
        description: Totalmarks
      negativemarks:
        type: string
        description: Negativemarks
      testtype:
        type: string
        description: TestType
      examtype:
        type: string
        description: ExamType
      isactive:
        type: string
        description: IsActive
      id:
        type: string
        description: id
      questionId:
        type: integer
        format: int64
        description: questionId
      correct:
        type: boolean
        description: question answer correct
      clearResponse:
        type: boolean
        description: clear Response
      markForReview:
        type: boolean
        description: mark for review
      response:
        type: string
        description: response
      assignmentQuestionId:
        type: integer
        format: int64
        description: assignmentQuestionId
      studentAssignmentId:
        type: integer
        format: int64
        description: studentAssignmentId
      isVisited:
        type: boolean
        description: isVisited
      questionGap:
        type: integer
        format: int64
        description: questionGap
      attemptedTime:
        type: string
        description: attemptedTime
      lmsNegativemarks:
        type: number
        format: "float"
        description: lmsNegativemarks
      lmsPostiveMarks:
        type: number
        format: "float"
        description: lmsPostiveMarks   
    xml:
      name: "sectionAssignmentTest" 
  assignmentTestPaper:
    type: "object"
    properties:
      sectionId:
         type: integer
         format: int64
         description: Section Id
      sectionName:
         type: string
         description: Section Name
      correct:
        type: "integer"
        format: "int32"
      Incorrect:
        type: "integer"
        format: "int32"
      Unattempted:
        type: "integer"
        format: "int32"
      positiveMarks:
        type: number
        format: "float"
      negativeMarks:
        type: number
        format: "float"
      AssignmentTestPaper:
        type: object
        items:
          type: object
          $ref: "#/definitions/sectionAssignmentTest"
    xml:
      name: "assignmentTestPaper"
...
  teacherQuestionLibraryInfo:
    type: "object"
    properties:
      questionId:
        type: "integer"
        format: "int32"
      text:
         type: "string"
      answer:
         type: "string"
      positiveMarks:
        type: "number"
        format: "float"
      negativeMarks:
        type: "number"
        format: "float"
      solution:
        type: "string"
      hint:
        type: "string"
      questionTypeId:
        type: "integer"
        format: "int32" 
      questionStatus:
        type: "string"
      subjectId:
        type: "integer"
        format: "int32"
      assignSectionId:
        type: "integer"
        format: "int32"
      questionOption:
        type: array
        items:
          type: "string"
      status:
        type: "boolean"
    xml:
      name: "teacherQuestionLibrary"
  getTeacherQuestionLibraryInfo:
    type: "object"
    properties:
      questionId:
        type: "integer"
        format: "int32"
      text:
         type: "string"      
      positiveMarks:
        type: "number"
        format: "float"
      negativeMarks:
        type: "number"
        format: "float"
      solution:
        type: "string"
      hint:
        type: "string"
      questionTypeId:
        type: "integer"
        format: "int32" 
      questionStatus:
        type: "string"
      subjectId:
        type: "integer"
        format: "int64"
      assignSectionId:
        type: "integer"
        format: "int64"
      questionOption:
          type: array
          items:
            $ref: "#/definitions/questionOptionInfo"      
    xml:
      name: "getTeacherQuestionLibraryInfo"
  questionOptionInfo:
    type: "object"
    properties:
        optionId:
          type: "integer"
          format: "int32"
        option:
          type: "string"
        answer:
          type: "boolean"
    xml:
      name: "questionOptionInfo"    
  Error:
    description: Schema for API errors.
    type: object
    properties:
      status:
        type: integer
        description: original HTTP error code, should be consistent with the response
          HTTP code
        minimum: 100
        maximum: 599
      code:
        type: string
        description: 'The error code, in screaming snake case: eg PAYLOAD_VALIDATION_ERROR'
        pattern: "^[A-Z]+(_[A-Z]+)*$"
      message:
        type: string
        description: descriptive error message for debugging
      moreInfo:
        type: string
        format: uri
        description: link to documentation to investigate further and finding support
      details:
        type: array
        description: list of problems causing this error
        items:
          title: errorDetail
          description: schema for specific error cause
          type: object
          properties:
            field:
              type: string
              description: JSON object notation specifying the element in the request
                payload causing the error. eg company.entitlements[3]
            code:
              type: string
              description: 'The error code detail in screaming snake case: eg MISSING_VALUE'
              pattern: "^[A-Z]+(_[A-Z]+)*$"
            message:
              type: string
              description: descriptive error detail message for debugging
            moreInfo:
              type: string
              format: uri
              description: link to documentation to investigate further and finding support
                for error detail
          required:
            - code
    required:
      - status
      - code